pyproject.toml
[tool.poetry]
name = "trekoserver"
version = "0.1.0"
description = ""
authors = ["shahzaibchenna"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-admin-dracula = "^0.0.3"
django-filter = "^24.3"
celery = {extras = ["redis"], version = "^5.4.0"}
boto3 = "^1.35.96"
django-cors-headers = "^4.7.0"
dotenv = "^0.9.9"
django-environ = "^0.12.0"
requests = "^2.32.5"


[tool.poetry.group.ai.dependencies]
face-recognition = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
